Add filter functionality.
Adds new class 'Filter' which represent objects that are passed every message to act on, rather than having to be invoked like commands. Command execution has been appropriately abstracted to the CommandExecutor filter.

YouTube Previewer is a new feature which parses messages for links to YouTube videos. If matched, the bot will send details about the video to the chat it was posted in.